Add lima-overlay service running before lima/cloud-init-local
It is used to modify files in the boot filesystem before they are
remapped by the Lima boot scripts to the data volume.

The previous approach using the "local" service doesn't work in an
upgrade because the service is not configured in the original ISO,
and since "local" runs last, the /etc directory has been replaced
by the previous version on the data volume, disabling the service
before it has a chance to run.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>