Update nerdctl 0.17.1 → 0.19.0
nerdctl 0.18.0+ no longer includes the CNI isolation plugin, so we'll
always install the cni-plugins now when LIMA_INSTALL_NERDCTL is set.

If cni-plugins 1.1.0 or later are installed, then nerdctl no longer
requires the isolation plugin.

Alpine 3.15 includes cni-plugins 1.0.1, so we need to install the
package from the edge repo.

Also in Alpine 3.15 the flannel plugin is no longer included in
cni-plugins, but not yet available as a separate package, so we have
to install cni-plugin-flannel from the edge repo as well. This should
have been done already when we moved from 3.14 to 3.15.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>